add alot of user routes

+++ TheHub/users/signals.py
@@ -6,11 +6,12 @@ from django.contrib.auth.models import User
 @receiver(post_save, sender=Invite)
 def create_friend(sender,instance, **kwargs):
     if instance.accepted:
-        friend = Friends.objects.create(inviter = instance.inviter, target = instance.target)
+        friend = Friends.objects.create(inviter = instance.inviter.profile, target = instance.target.profile)
         friend.save()
         instance.delete()
-    if instance.accepted == False:
-        instance.delete()
+    if instance.accepted is not None:
+        if instance.accepted == False:
+                instance.delete()
 
 @receiver(post_save, sender=User)
 def create_profile(sender, instance, created, **kwargs):
